Move graphical primitive classes out of PrimitiveMan
